Frogger: Move VisualElement class into a separate game utility file.
