+++ app/game/frogger/Frogger.py
@@ -1,32 +1,5 @@
 from app.utilities.Colours import Colours
-
-class VisualElement:
-    def __init__(self, colour, x_length, y_length, x_start, y_start):
-        self.colour = colour
-        self.x_length = x_length
-        self.y_length = y_length
-        self.x = x_start
-        self.y = y_start
-
-        self.x_start = x_start
-        self.y_start = y_start
-        self.colour_start = colour
-
-    def get_render_properties(self):
-        return [((self.x, self.y), (self.x_length, self.y_length)), self.colour]
-
-    def get_footprint(self):
-        footprint = [self.x, self.x + self.x_length - 1, self.y, self.y + self.y_length - 1]
-        return footprint
-
-    def change_colour(self, colour):
-        self.colour = colour
-
-    def reset(self):
-        self.x = self.x_start
-        self.y = self.y_start
-        self.colour = self.colour_start
-
+from app.game.utilities.VisualElement import VisualElement
 
 class Frog(VisualElement):
     def __init__(self, colour, x_length, y_length, x_start, y_start, x_limit, y_limit):
@@ -44,7 +17,6 @@ class Frog(VisualElement):
 
         if new_y >= 0 and new_y <=self.y_limit:
             self.y = new_y
-
 
 class Enemy(VisualElement):
     def __init__(self, colour, x_length, y_length, x_start, y_start, y_limit, direction, velocity):
@@ -70,7 +42,6 @@ class Enemy(VisualElement):
 
         if self.direction == 'up' and self.y < -self.y_length:
             self.y = self.y_start
-
 
 class FroggerGame():
     def __init__(self, x_boundary, y_boundary):
